Add skeleton to switch between different analysis modes
